change mcbm key to uint64_t type and fix bug by implementing key serializer for 64

+++ benchmark/mcbm/table_decl.h
@@ -11,7 +11,7 @@
 
 namespace sql {
 
-FIELD(uint32_t, k);
+FIELD(uint64_t, k);
 KEYS(McbmKey);
